// src/app/article-page-edit/types/edit-article-state.interface.ts
import { ArticleInterface } from 'src/app/article-page-show/types/article.interface'
import { BackendErrorsInterface } from 'src/app/shared/types/backend-errors.interface'

export interface EditArticleStateInterface {
  article: ArticleInterface | null
  isLoading: boolean
  isSubmitting: boolean
  validationErrors: BackendErrorsInterface | null
}